Moving FormControls to src/ui
