+++ src/ui/FormControls/MenuButton.cs
@@ -1,0 +1,57 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Windows.Forms;
+using System.ComponentModel;
+using System.Drawing;
+
+namespace SM64DSe.FormControls
+{
+    // Taken from: https://stackoverflow.com/questions/10803184/windows-forms-button-with-drop-down-menu#24087828
+
+    public class MenuButton : Button
+    {
+        [DefaultValue(null)]
+        public ContextMenuStrip Menu { get; set; }
+
+        [DefaultValue(false)]
+        public bool ShowMenuUnderCursor { get; set; }
+
+        protected override void OnMouseDown(MouseEventArgs mevent)
+        {
+            base.OnMouseDown(mevent);
+
+            if (Menu != null && mevent.Button == MouseButtons.Left)
+            {
+                Point menuLocation;
+
+                if (ShowMenuUnderCursor)
+                {
+                    menuLocation = mevent.Location;
+                }
+                else
+                {
+                    menuLocation = new Point(0, Height);
+                }
+
+                Menu.Show(this, menuLocation);
+            }
+        }
+
+        protected override void OnPaint(PaintEventArgs pevent)
+        {
+            base.OnPaint(pevent);
+
+            if (Menu != null)
+            {
+                int arrowX = ClientRectangle.Width - 14;
+                int arrowY = ClientRectangle.Height / 2 - 1;
+
+                Brush brush = Enabled ? SystemBrushes.ControlText : SystemBrushes.ButtonShadow;
+                Point[] arrows = new Point[] { new Point(arrowX, arrowY), new Point(arrowX + 7, arrowY), new Point(arrowX + 3, arrowY + 4) };
+                pevent.Graphics.FillPolygon(brush, arrows);
+            }
+        }
+    }
+}
